chore(docs): Optimized array indexing in the log monitor widget's rendering logic
# Project State

## Current Focus
Optimized array indexing in the log monitor widget's rendering logic

## Context
The change improves performance by removing unnecessary casting in the array index calculation during status bar rendering.

## Completed
- [x] Removed redundant `as usize` cast in array index calculation
- [x] Simplified index calculation for better readability and potential performance

## In Progress
- [x] Performance optimization of the log monitor widget

## Blockers
- None identified

## Next Steps
1. Verify no visual rendering issues after the change
2. Consider additional performance optimizations in the log monitor widget
---
category: chore
scope: docs
---

+++ examples/_cookbook/log_monitor.rs
@@ -256,7 +256,7 @@ impl Widget for LogMonitor {
         let auto_str = if self.auto_scroll { "auto" } else { "paused" };
         let status = format!(" 󰔱 {}  |  󰑎 {} lines  |  scroll: {}  |  c=clear r=resume  |  q=quit", last_str, self.total_lines, auto_str);
         for (i, c) in status.chars().enumerate().take(w - 2) {
-            let idx = ((h - 1) * w + 1 + i) as usize;
+            let idx = (h - 1) * w + 1 + i;
             if idx < p.cells.len() {
                 p.cells[idx] = Cell { char: c, fg: t.fg_muted, bg: t.surface_elevated, style: Styles::empty(), transparent: false, skip: false };
             }
